// src/enums/automations.ts
/**
 * Enum for account type values.
 * @readonly
 * @enum {strings}
 */
export enum Automations {
  TAG_COMMENT_CLIP = 'TAG_COMMENT_CLIP',
  MOVE_COPY = 'MOVE_COPY',
  EDIT_TRANSCRIPT_TEXT = 'EDIT_TRANSCRIPT_TEXT',
  EDIT_TRANSCRIPT_TYPE = 'EDIT_TRANSCRIPT_TYPE',
  CREATE_LABEL = 'CREATE_LABEL',
}